# swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agrirouter Middleware",
    "description": "Middleware to access the agrirouter using a 'smaller' API. The middleware will handle the connection for all endpoints and fetches / stores messages from the agrirouter within a custom database.",
    "contact": {
      "name": "Agrirouter Middleware",
      "url": "https://github.com/agrirouter-middleware"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://github.com/agrirouter-middleware/LICENSE"
    },
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/secured/api/non-telemetry-data/publish/video/{externalEndpointId}": {
      "post": {
        "tags": [
          "non telemetry data management"
        ],
        "description": "Publish video files for an existing endpoint.",
        "operationId": "non-telemetry-data.publish-video",
        "parameters": [
          {
            "name": "externalEndpointId",
            "in": "path",
            "description": "The external endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishVideoDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Response if the message was published successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
    "/secured/api/non-telemetry-data/publish/taskdata/{externalEndpointId}": {
      "post": {
        "tags": [
          "non telemetry data management"
        ],
        "description": "Publish a task data file for an existing endpoint.",
        "operationId": "non-telemetry-data.publish-taskdata",
        "parameters": [
          {
            "name": "externalEndpointId",
            "in": "path",
            "description": "The external endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishNonTelemetryDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Response if the message was published successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
    "/secured/api/non-telemetry-data/publish/shape/{externalEndpointId}": {
      "post": {
        "tags": [
          "non telemetry data management"
        ],
        "description": "Publish a shape file for an existing endpoint.",
        "operationId": "non-telemetry-data.publish-shape",
        "parameters": [
          {
            "name": "externalEndpointId",
            "in": "path",
            "description": "The external endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishNonTelemetryDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Response if the message was published successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
    "/secured/api/non-telemetry-data/publish/pdf/{externalEndpointId}": {
      "post": {
        "tags": [
          "non telemetry data management"
        ],
        "description": "Publish PDF files for an existing endpoint.",
        "operationId": "non-telemetry-data.publish-PDF",
        "parameters": [
          {
            "name": "externalEndpointId",
            "in": "path",
            "description": "The external endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishNonTelemetryDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Response if the message was published successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
    "/secured/api/non-telemetry-data/publish/image/{externalEndpointId}": {
      "post": {
        "tags": [
          "non telemetry data management"
        ],
        "description": "Publish image files for an existing endpoint.",
        "operationId": "non-telemetry-data.publish-image",
        "parameters": [
          {
            "name": "externalEndpointId",
            "in": "path",
            "description": "The external endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishImageDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Response if the message was published successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "The timestamp (UTC) of the response.",
            "format": "date-time"
          },
          "key": {
            "type": "string",
            "description": "The key of the error."
          },
          "errorMessage": {
            "type": "string",
            "description": "The message of the error."
          }
        },
        "description": "Generic error response to map the internal errors and provide clean API description."
      },
...
      "PublishNonTelemetryDataMessageTuple": {
        "required": [
          "fileName",
          "messageContent"
        ],
        "type": "object",
        "properties": {
          "messageContent": {
            "type": "string",
            "description": "The Base64 encoded message content."
          },
          "fileName": {
            "type": "string",
            "description": "The name of the file."
          },
          "recipients": {
            "type": "array",
            "description": "The recipients for direct sending. If there is a recipient, the file will not be published.",
            "items": {
              "type": "string",
              "description": "The recipients for direct sending. If there is a recipient, the file will not be published."
            }
          }
        },
        "description": "Internal class to encapsulate the message tuples."
      },
      "PublishVideoDataRequest": {
        "required": [
          "messageTuples",
          "videoType"
        ],
        "type": "object",
        "properties": {
          "messageTuples": {
            "type": "array",
            "description": "The message tuples.",
            "items": {
              "$ref": "#/components/schemas/PublishNonTelemetryDataMessageTuple"
            }
          },
          "videoType": {
            "type": "string",
            "description": "The video type, used to map the specific agrirouter message type.",
            "enum": [
              "AVI",
              "MP4",
              "WMV"
            ]
          }
        },
        "description": "The request body containing all necessary information to publish video files."
      },
      "PublishNonTelemetryDataRequest": {
        "required": [
          "messageTuples"
        ],
        "type": "object",
        "properties": {
          "messageTuples": {
            "type": "array",
            "description": "The message tuples.",
            "items": {
              "$ref": "#/components/schemas/PublishNonTelemetryDataMessageTuple"
            }
          }
        },
        "description": "The request body containing all necessary information to publish PDF files."
      },
      "PublishImageDataRequest": {
        "required": [
          "imageType",
          "messageTuples"
        ],
        "type": "object",
        "properties": {
          "messageTuples": {
            "type": "array",
            "description": "The message tuples.",
            "items": {
              "$ref": "#/components/schemas/PublishNonTelemetryDataMessageTuple"
            }
          },
          "imageType": {
            "type": "string",
            "description": "The image type, used to map the specific agrirouter message type.",
            "enum": [
              "JPEG",
              "PNG",
              "BMP"
            ]
          }
        },
        "description": "The request body containing all necessary information to publish image files."
      },
...
    },
    "securitySchemes": {
      "secured": {
        "type": "http",
        "description": "Default security scheme, used for the private flows.",
        "scheme": "basic"
      }
    }
  }
}